requirements.txt
aiogram
aiohttp_socks
requests
tcping
python-dotenv
